pyproject.toml
[project]
name = "stormglass"
version = "0.1.0"
dependencies = [
    "arrow~=1.0",
    "requests~=2.0",
    "python-dotenv~=1.0",
]
